+++ packages/disco/lib/src/disco_internal.dart
@@ -4,11 +4,9 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
 part 'models/overrides/override.dart';
-part 'models/overrides/provider_argument_override.dart';
-part 'models/overrides/provider_override.dart';
 part 'models/providers/instantiable_provider.dart';
-part 'models/providers/provider_argument.dart';
-part 'models/providers/provider.dart';
+part 'models/providers/arg_provider.dart';
+part 'models/providers/no_arg_provider.dart';
 part 'utils/disco_config.dart';
 part 'utils/extensions.dart';
 part 'widgets/provider_scope.dart';