+++ packages/disco/lib/src/models/overrides/override.dart
@@ -5,3 +5,35 @@ part of '../../disco_internal.dart';
 sealed class Override {
   Override._();
 }
+
+/// Override that, if inserted into the widget tree, takes precedence over
+/// [_originalArgProvider].
+@immutable
+class ArgProviderOverride<T extends Object, A> extends Override {
+  ArgProviderOverride._withArgProvider(
+    this._originalArgProvider,
+    this._mockArgProvider,
+  ) : super._();
+
+  /// The reference of the argument provider to override.
+  final ArgProvider<T, A> _originalArgProvider;
+
+  /// The reference of the argument provider override.
+  final ArgProvider<T, A> _mockArgProvider;
+}
+
+/// Override that, if inserted into the widget tree, takes precedence over
+/// [_originalProvider].
+@immutable
+class ProviderOverride<T extends Object> extends Override {
+  ProviderOverride._withProvider(
+    this._originalProvider,
+    this._mockProvider,
+  ) : super._();
+
+  /// The reference of the provider to override.
+  final Provider<T> _originalProvider;
+
+  /// The reference of the provider override.
+  final Provider<T> _mockProvider;
+}